[BUG FIX] #439: color int 크기 초과 버그 해결

+++ backend/src/main/java/autoever2/cartag/service/ColorService.java
@@ -66,7 +66,7 @@ public class ColorService {
         if (count.isEmpty()) {
             return 0;
         }
-        return boughtCount.intValue() * 100 / count.get().intValue();
+        return (int) (boughtCount * 100 / count.get());
     }
 
     public String changeUrl(String value, int index) {
